Create README.md
: Smart Maintenance Portal is the user based web application which is used to
solve the problems in our day to day life or problems in society such as water related
issue, street light issue, road related issue, etc. specially the problems which comes
under the Government. this problems don’t get solved easily and it takes much time to
solve the problems in physical way. So to get the fast solution of the problem we created
this web application called as Smart Maintenance Portal. This web Application consists
of three panels Admin, Department and User. In Admin Panel the admin adds the
different types of department like Road, Water, Electricity, Street light, etc. In
department panel the specific department has assign the task or problem registered by
the user and has to solve the problem registered by the user. In User panel the user
registers the complaint faced by him in his society or any area and can register the
complaint according to the specific department. Rather than this the application consists
extra features also. This all the problems comes under the Government and this
problems can be solved fast and in efficient way